Ensure all values are emitted before raising
The idea is to have a stream that emits all valid values up to the first
exception. Discarding the currently accumulating chunk can result in
some data being sometimes emitted or not, depending on the chunk size.
Observable behavior of the stream should not depend on how data is
chunked, so all valid data _so far_ must always be emitted.