Cache the `byHeader` field
Having it as a def allowed for statically checking that headers are
defined and avoid accidental call to it when no headers are defined.
However, this implies that the map is recomputed for each access to a
function like `as`, which induces significant overhead.

Turning it into a lazy has these implications:
 - significantly improve performance when transforming rows with lots of
   cells
 - calling this field becomes more unsafe since there is no static check
   anymore that the headers are actually defined

However the risk is mitigated by the fact that this private field is
called by a quite stable API which has the constraint, ensuring the call
is safe and by adding a deprecation warning so that new uses will at
least be made aware of having to check things.