Add a way to set a value that adds missing column
This new function adds the missing header in case it was not present in
the original row.
Due to potential problem implementing a correct behavior for index based
setting in case of rows with headers, this function is only added for
name based access.
A new function is added, not to break existing code with the former
behavior of ignoring missing headers.