Add support for codecs
In addition to parsing and transforming Json ASTs, this adds a way to
work with plain Scala value when the underlying Json library has some
concept of codecs.

This is also the oppportunity to rationalize the selector code a bit,
making the trasformation pipes more regular and generic.

Old `json` package level transform pipes are deprecated in favor of
their variant in the `ast` package. This allows for semantically scoped
transformation pipes.