Add copyless streaming tree transducers
The copyless natures is ensured by using a controlled environment update
language, without losing expressiveness.

This execution model emits as soon as possible, and requires transducers
to be total, otherwise the stream fails.
This also ensures that everything is emitted up to the first error.

The STT model can be used to implement query languages on tree
structures and some transformations.