Allow for XML postlog
The specification allows for `misc` appearing after the root element,
and the event parser should support this, as well as the document
builder.

Since a stream can contain concatenated XML documents, a choice has to
be made regarding how misc is associated between 2 documents in the
stream. This implementation makes the choice to associate the misc as
postlog of the previous document since it allows to emit early without
knowing what comes next (i.e. whether the stream actually contains only
one document or several documents).