Add streaming pretty printer utilities
It is inspired by the paper _Lazy v. Yield: Incremental, Linear
Pretty-printing_ by Oleg Kiselyov, Simon Peyton-Jones, and Amr Sabry.
It adds indentation and alignment features to be useful in the context
of streaming tree structure pretty printing (XML, JSON, ...).

Any type that has an instance of `Renderable` can be printed.

The printer guarantees that the data is printed as early as possible
and the it is correct up to the first malformed tree structure.