Refine msgpack fixpoint test
The parser mapping ByteVector to MsgpackItem can be seen as a not
injective morphism, that is, there are many ByteVectors that will map to
the same MsgpackItem. Because of this, we cannot possibly guarantee that
`serialize(parse(bs))` is fixpoint for an arbitrary `bs`. However,
currently implemented serializers *are* injective (if we exclude the
Timestamp format family as it can be represented with Extension types)
and so, we can guarantee `serialize(parse(bs)) == bs` if `bs` is a
member of a subset of ByteVector that is emitted by a serializer.

In other words, the following code will be true for any `bs` if
`serialize` is injective and we ignore the Timestamp type family:
```
 val first = serialize(parse(bs))
 val second = serialize(parse(first))
 first == second
```

This test makes sure that the above holds.